+++ model/controllers/operations/getById.js
@@ -1,0 +1,8 @@
+import contacts from "../contacts.json";
+
+const getContactById = async (contactId) => {
+  const [contact] = contacts.filter((contact) => contact.id === contactId);
+  return contact;
+};
+
+export default getContactById;
