+++ model/controllers/operations/index.js
@@ -1,0 +1,13 @@
+import addContact from "./addContacts";
+import getContactById from "./getById";
+import listContacts from "./listContacts";
+import removeContact from "./removeContacts";
+import updateContact from "./updateContacts";
+
+export default {
+  listContacts,
+  getContactById,
+  removeContact,
+  addContact,
+  updateContact,
+};
