+++ model/controllers/operations/listContacts.js
@@ -1,0 +1,7 @@
+import contacts from "../contacts.json";
+
+const listContacts = async () => {
+  return contacts;
+};
+
+export default listContacts;
